website/pyproject.toml
[project]
name = "cross-docs-website"
version = "0.1.0"
description = "Cross-Docs documentation website"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "cross-inertia>=0.10.0",
    "cross-docs",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "watchfiles>=0.20.0",
]

[tool.uv.sources]
cross-docs = { path = "../python", editable = true }
